\name{mlma-package}
\alias{mlma-package}
\docType{package}
\title{
Multilevel Mediation Analysis
}
\description{
The package is used to do mediation analysis with generalized multilevel models.
}
\details{
\tabular{ll}{
Package: \tab mlma\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-28\cr
License: \tab GPL (>= 2)\cr
}
\code{"\link[=data.org]{data.org}"} is used to transform the variables and organize the predictor, mediators and outcome into the format that are ready to be used for multilevel mediation analysis. \code{"\link[=mlma]{mlma}"} is for multilevel mediation analysis on the original data set. \code{"\link[=boot.mlma]{boot.mlma}"} is a combined function that organized data set, do multilevel mediation analysis on original data sets and bootstrapping samples.}
\author{
Qingzhao Yu  \email{qyu@lsuhsc.edu}, Bin Li \email{bli@lsu.edu}

Maintainer: Qingzhao Yu  \email{qyu@lsuhsc.edu}
}
